Disable NllWithSoftmaxGradDirect pattern if the fwd and bwd Ops are on different pipeline stages
Summary:
An alternative solution to T28759.

D34329 Moved the Op to the fwd pipelineStage which satisfies scheduling constraints for Ops depending on the "loss" output. However, this would break scheduling constraints for Ops that produce the "getGradProbsInIndex" input. While you could propagate the pipelineStage change to those producer Ops, I think that would be a overkill.

The impact of disabling the pattern will be a small amount extra computation in the fwd stage.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Maniphest Tasks: T28759

Differential Revision: https://phabricator.sourcevertex.net/D34472