Devicex refactor to move serializable state into separate class
Summary:
Refactor Devicex so that objects and functions involved in `poplar::Graph` creation are kept separate from objects and functions used in `poplar::Engine` calls.
This is a first step to be able to serialize state that is used for lowering the popart IR to Poplar.

The objects and functions previously used for `poplar::Graph` creation are moved to a separate class `popart::popx::IrLowering`. I'm not set on the name `IrLowering`.
`Devicex` has an accessor for the lowering state such that ops that make use of `Devicex` functions and state can access as before through the accessor. Functionality related to running
such as: `stepio`, connecting streams to buffers/callbacks, etc. are retained in `Devicex`

I've removed some of the public API in `Devicex` which may risk breaking some custom ops, although I did ensure that the BERT custom ops still compile.

Test Plan: All tests pass.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T28409

Differential Revision: https://phabricator.sourcevertex.net/D34889