Convert (Flatten,Squeeze,Unsqueeze) to Reshape earlier.
Summary:
The only complication faced was that input ids were sometimes unavailable.
A new method was added to OpManager to solve this problem.

Motivation for this diff came up when adding a new virtual method to the Op
class, and needing to duplicate the solution across all 4 variants of Reshape.

One solution would have been inheritance, but this diff's approach seems cleaner to
me, and ~800 lines of code are removed. A user can still use use all 4 variants of Reshape,
and they are still available in the TransformationBuilder.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, matthewha

Subscribers: matthewha

Maniphest Tasks: T29498

Differential Revision: https://phabricator.sourcevertex.net/D35307