Support pipelining when a tensor is consumed by ops on more than two pipeline stages
Summary:
Before:
With and without RecomputationType::Pipeline, if a to-stash tensor must be restored on >1 different pipeline stages, error.

After:
If RecomputationType::Pipeline, if a to-stash tensor must be restored on >1 different pipeline stages for direct consumption, supported, but if it must be restored on >1 different pipeline stages to recompute a descendent tensor, error (to be supported in T30014)
Else, if a to-stash tensor must be restored on >1 different pipeline stages, supported.

Fixes T29611

Test Plan: A new test has been added, tests/popart/pipelining_tests/pipeline_multiple_restores_per_stash.py::test_2_restores_for_1_stash

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, george, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, matthewha

Subscribers: matthewha, oskarb, jamesn

Maniphest Tasks: T29611

Differential Revision: https://phabricator.sourcevertex.net/D35810