Serialization of PopART IR and lowered IR structures
Summary:
  - Add capnproto dependency to popart build
  - Create capnp schemas for PopART structures needed to load and save a lowered PopART IR.
  - Implemented serialization and deserialization and added a unit test.
  - Introduction of a new PopART abstraction `popx::Executablex` which is meant to be a similar abstraction to `poplar::Executable`, i.e `popx::Devicex` loads and runs a `popx::Executablex` which can be serialized/deserialized.
  - Some refactoring to existing structures to implement serialization

Halfway through implementing IR serialization I felt that it was unnecessary to flesh out everything related to SessionOptions, DataFlow, anchors etc. since they will be passed to the `popart::Session` anyway.
We can just use a hash to make sure that they are the same and if there is other stuff from the IR that needs to be serialized then we can add it later on.

Test Plan:
All tests pass, new test added.

The test that's supposed to test the replicated weight sharding serialization is disabled since I haven't been able to run it.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, fabiant, mohammeds

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T29564

Differential Revision: https://phabricator.sourcevertex.net/D36124