Fix false-positive warnings of IrLowering::verifyTaskOrder, now thow error on failure
Summary:
I changed the warning to an error, to see what would fail, and saw the following:
```
	280 - IpuModel_default_pipeline_quadratic_epsilon_test_1 (Failed)
	281 - IpuModel_default_pipeline_recompute_numerical_test_0 (Failed)
	282 - IpuModel_default_pipeline_anchor_recomputed_tensor_check (Failed)
	287 - IpuModel_default_pipeline_full_recompute_test (Failed)
	290 - IpuModel_default_pipeline_test (Failed)
	293 - IpuModel_default_pipeline_multiple_restores_per_stash (Failed)
```
In all cases, these warnings were false positives. The were all of the case:
- Iterating through the tasks in their final order, you come across `task_OpA`
- Iterating through the ops that must be recomputed to produce the inputs to `OpA` (in order first to last), you come across `OpB`
- Iterating through the topocons of `OpB`, you come across `OpC`
- You check whether `OpC` has been seen twice
  - once for the first time it appears in the graph
  - and once for when is run to for the recomputation of `OpA`.
- `OpC` fails this check.

But this is too strict. Imagine a graph like:
```
  checkpoint_tensor -- OpB -- t0 -- OpA -- t1
                               \
                              OpC
```
where you have a topocon from `OpC` to `OpB`. `OpC` isn't needed for the recomputation of `OpA`'s inputs, so it shouldn't fail this check.

Instead, only check `OpC` has been seen twice if it is also required for the recomputation of `OpA`.

All the above tests now pass (with verifyTaskOrder throwing an error if it catches anything).

Fixes T29166

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, markk, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, jamesn

Subscribers: jamesn, matejk

Maniphest Tasks: T29166

Differential Revision: https://phabricator.sourcevertex.net/D36613