Prune redundant identityloss ops
Summary:
Poptorch training models always apply an identityloss to the user-provided loss, even if scalar.

In the case of identityloss' input being a scalar, it can be replaced by an identity op (regardless of reduction type). This is a minor optimization - it saves the copying of two scalar tensors - but the main motivation was to fix the root cause of T30749

Fixes T30887

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb, george

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb, george

Maniphest Tasks: T30887

Differential Revision: https://phabricator.sourcevertex.net/D36930