In `popart-config.cmake`, have the targets actually find and link their required dependencies
Summary:
Depends on D36623.

**Before**

`willow/CMakeLists.txt` creates `popart-only` library targets and `popart`, `popart-internal` interface targets that expose dependencies to increasing degrees of visibility.
`popart-only` was only popart, no poplar, poplibs, gcl, libpvti. However, we did publicly include poplar, as users must have its headers to build popart. Similarly then, the config file did this too.

This hackery is required, fundamentally, because Popart does not enforce a clear separation between public and private headers, and thus does not hide dependencies.
However, Poptorch uses Popart headers such that ONNX, Protobuf, Poplar SDK are not needed to build, and wanted a popart target that does not have them publicly linked/included. Thus, we created these modular targets.

The config file, for `popart-only`, was well-defined, as it found all its dependencies (which are none). However, for `popart`, it was ill-defined as it did not find any of its dependencies.

In order to install the `popart` target, but not have any of the interface includes/libs (because the config file isn't finding the deps), we wrap the `target_include_directories` etc. arguments in willow with `<BUILD_INTERFACE:...>` genexes.

We print a warning in the config file saying the `popart` target is ill-defined.

**After**

 We ship two separate `COMPONENTS`, `popart-only` and `popart`.  The default is `popart`.

Again, due to the lack of public-private separation in Popart, it is un-maintainable to try to hide Poplar SDK headers from the Popart headers that Poptorch includes. Thus, we moved those libs to be public in `popart-only`. However, ONNX and Protobuf are still public only in `popart`.

To implement the above, `willow/CMakelists.txt` now installs two separate components.

The config file handles the requesting of components and defaults to `popart`.

The config file finds the required dependencies of the requested components only. Thus, if only `popart-only` is requested, the user does not need to provide ONNX or Protobuf.

As poplar (and the whole SDK) is public in popart-only, we remove the `target_include_directories(popart-only PUBLIC poplar)` code in both willow and the config file.

Thanks to previous subtasks of T28375 making popart and how we find Popart's dependencies all target-based (instead of variable-based), and the fact that the config file now finds the required dependencies; we remove the `<BUILD_INTERFACE:...>` genexes in willow.

We install Popart's `Findgcl.cmake` for the user as gcl does not provide a config module.

**Other bits of cleanup/implementation details:**

In the config file, remove the warning file about `popart` being ill-defined.

In the config file, remove if NOT TARGET condition around include of targets files as they already handle repeat inclusions.

Fix typo in comment dyanamic -> dynamic.

Fix `POPART_CMAKE_INSTALL_CONFIGDIR` cache doc string.

In config file, use `_safe_find_dependency` that preserves the value of `PACKAGE_PREFIX_DIR`.

~~In config file, properly set `POPART_INCLUDES` and `POPART_LIBRARIES`, so they include the dependencies of all components requested. This is only one level deep: if dependencies' dependencies are required, the user can manually get them from what is provided in these variables.~~

Remove `POPART_INCLUDES` and `POPART_LIBRARIES`, because:
      1. Setting them adds a lot of complexity.
      2. Are they even that useful? They may not be exactly what the user needs, or not enough.
      3. For what the user does need, they can manually recover from the targets as I have here.
      4. At two points in the implementation, there are things that are slightly high-maintenance
I have saved the implementation on another branch. If someone requests something like them, we can bring them back. Poptorch is the only user of the config file and they do not use these variables, so removing them should be fine.

Update config file documentation accordingly.

In willow, extract common includes that are private in `popart-only` but public in `popart-internal` (that is, they are generally private, but we expose them to internal targets like tests) into a variable `popart_common_includes`.

Test Plan: Test that Poptorch (with D36623) still builds before and after landing this diff, and without extra warnings. Both in the view and standalone.

Reviewers: anthonyb, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Subscribers: jamesn

Maniphest Tasks: T29285

Differential Revision: https://phabricator.sourcevertex.net/D36856