Upload PopART supported ops file as an artifact
Summary:
What this is *intended* to do is upload the generated `suported_ops_gen.md` file to Artifactory (and/or SWDB) so it can be used to build the externally published docs. It shouldn't be included in the package used to build the SDK (although it doesn't really matter if it is).

I'm not really sure if I have understood how the cmake files work, and therefore if I have done the right thing or not.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Subscribers: jamesn

Maniphest Tasks: T28480

Differential Revision: https://phabricator.sourcevertex.net/D37692