Forward cache variables correctly in `SuperProjectConfig.cmake`
Summary:
Remove PIC cache var in SuperProjectConfig, this does not even set it for popart,
but for CBT. We do not need a cache var for this as we don't need it as a user option.
Popart CMake already sets this appropriately.

Helper function in SuperProjectConfig.cmake for implicitly passing through vars
to POPART_CMAKE_ARGS.

Move all cache vars to Popart. Remove "if in view" logic when setting
POPART_CBT_VIEW_DIR in SuperProjectConfig, as must always be in view.

Only log values of POPART_PKG_DIR and POPART_CBT_VIEW_DIR when defaulted, as
otherwise is not useful. POPART_CBT_VIEW_DIR is VERBOSE log level because it's
an internal variable that will thus get defaulted whenever the user builds,
but they don't care about it.

Remove unused GNUInstallDirs include from SuperProjectConfig. This does not
include it in popart anyway, only the cbt superproject.

Test Plan: Manual testing of vars.

Reviewers: samk, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: samk, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Maniphest Tasks: T32416

Differential Revision: https://phabricator.sourcevertex.net/D38497