Using mkdoc.py to copy hpp comments to python
Summary:
"Final" update (before any last corrections):

Idea is to run the `gen_python_docs.sh` file which runs pybind11_mkdoc module (I had the best results using the module, not the fle in the pybind11 repo) to copy the hpp comments into variables, which are the used in popart.cpp as python comments. This should be run every time there is a significant update to the hpp docs. Alternatively, minor changes can be updated manually, though these will be overwritten once the `gen_python_docs.sh` is rerun. To access the docs use the macro `DOC(popart, Subclass, Function)` as the last argument in the binding. Further explanation in the sh file and on the github repo for pybind11_mkdoc . I have added the DOC macro to add docs to a fair few of the popart.cpp functions, but more can be added in future.

There is also the file `willow/include/popart/pydocs_popart_custom.hpp` which is manually created and updated. This is for custom python docs you do not want to mirror the hpp docs. A similar `DOC(custom, Subclass, Function)` is used to access these.

There are also `willow/include/popart/docs/opset9_docs.hpp` for each opset, which s generated as part of the `gen_operators.py` run. this just copies the docs from the onnx spec into a variable, then references it in the generated cpp binding file. This is run as part of `gen_operators.py`, so doesn't need any extra effort. This could probably be improved to make the docs a bit better in future as it's just a description of the operator at present.

Test Plan: Tested locally and works. Since this is not run as part of the build, it should be relatively safe to change.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb, evan, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb, evan, jamesn

Subscribers: evan, grahamh, jamiep, hatemh, jamesn, shirazb, markk

Maniphest Tasks: T11818

Differential Revision: https://phabricator.sourcevertex.net/D33970