Enabled loading OfflineIpu executables into Ipu device
Summary:
Executables are compiled, and serialized with configuration of the
current run-time environment. When an executable's configuration doesn't
match with that of the run-time environment that it is loaded into, the
executable cannot be used to run pre-compiled model. However, an
`OfflineIpu` device type can be assumed equivalent to an `Ipu` device,
given that their respective configurations match. Therefore, executables
that are compiled on an `OfflineIpu` can be used on `Ipu` devices.

Reviewers: shauryas, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: shauryas, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T31785

Differential Revision: https://phabricator.sourcevertex.net/D38361