Lowering CallOps into multiple fragments (as determined by LivenessAnalyzer).
Summary:
This diff introduces a notion of 'subgraph parts'  -- meaning a fragment/sequence/function in which ops are lowered. Now, subgraphs are not always loaded into a single such part -- they are potentially lowered over multiple.

Subgraphs are split into multiple parts whenever a parent graph does an input/output copy somewhere inside a call to the subgraph. This is determined by the `SubgraphPartitioner` class. This class basically reads the results from the LivenessAnalyzer and uses this to determine how subgraphs are to be partitioned.

Most ops fall into one part of the subgraph they reside in. The only exception is `CallOps`. A CallOp's lowering essentially comprises copying of inputs/ calling subgraph parts of the called subgraph / copying of outputs. It may be that the best place for a parent subgraph to copy one of its inputs or outputs is somewhere in the middle of this sequence. To facilitate this, whenever this is the case, CallOps are lowered over multiple parts.

Motivation:

* The cost of calling poplar functions is assumed to be low (akin to a `jmp` or `call` in `x86`) so breaking up subgraphs is not thought to be expensive.
* This change allows us to copy inputs later / copy outputs earlier in future, potentially significantly (?) reducing tensor liveness when outlining.

Notes:

* `popprograms` now facilitates multiple fragments per subgraph.
* `irlowering` now allows lowering ops over multiple poplar sequences.
* At the moment the LivenessAnalyzer will always copy inputs, call the subgraph, copy outputs -- in that order. For this reason, at present, it should never be the case that a parent does a copy in the middle of the called subgraph. This means that at present each subgraph should comprise exactly 1 part -- e.g. no functional changes.

Reviewers: fabiant, evan, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Reviewed By: fabiant, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Subscribers: simonl, fabiant

Maniphest Tasks: T30551

Differential Revision: https://phabricator.sourcevertex.net/D38080