Add HistogramOp and Opx and unit test
Summary:
First part of T23646. The histogram op takes an input tensor, and a vector of 'bin edges' (via the `levels` attribute), and outputs a tensor with the counts of the number of elements in the input tensor in each bin.

It implements popop's `popops::histogram`.

A fairly straightforward op to implement. There's one design decision that has to be made, because `levels` is passed to popops as a poplar tensor. So you have to either:
1. Add the `levels` attribute as a constant tensor in the IR, and connect as a second input to the HistogramOp, or
2. Keep `levels` as an attribute (and append to extra outline attributes), and create a constant poplar tensor in HistogramOpx::grow.

I can't think of why either are obviously better. I think (2) is less code, so I went with that.

Fixes T23836

Test Plan: Added histogram_test.cpp

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb, george

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Subscribers: hatemh, jamesn, matthewha

Maniphest Tasks: T23836

Differential Revision: https://phabricator.sourcevertex.net/D39108