Do not scale by replication factor in the IdentityLossGradOps added by the batch serialization transform
Summary:
There is an issue in popart to do with the interaction between batch serialization and graph replication, when the loss has reduction-type 'mean'. The scaling of the replica losses (to give the effect of averaging over replicas) is currently done within the grow functions of the loss grad ops. The batch serialization transform adds an extra identityloss op when reducing the sub-mini-batch outputs. This reduction is scaling by 1/replication_factor, which is incorrect. This scaling should only be applied by an op that is the final loss in the graph (i.e. the source of backpropagation).

T33184 is to fix this issue across all loss ops (identity, l1, nll).

Fixes T30671
Also fixes one of the issues seen in T33274

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, jamesn, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, matthewha

Subscribers: matthewha

Maniphest Tasks: T30671

Differential Revision: https://phabricator.sourcevertex.net/D39611