Add more types and functions to the API doc
Summary:
Added more classes that need documenting. This includes base classes of currently documented classes (e.g. Session), types that are used as function parameters and classes that are referenced elsewhere in the documentation (e.g. Op and the related classes mentioned in the section on creating custom ops).

A lot of the added things are not well documented themselves. I will submit further diffs for those.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, simonl, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Maniphest Tasks: T17979

Differential Revision: https://phabricator.sourcevertex.net/D40045