Add another inplacing condition for pipelining
Summary:
PipelineCycles are not represented explicitly in the IR. As such certain
cases of inplace conflict must be handled here.
A pipeline cycle with recomputation can cause tensors to be overwritten.
On a single IPU a pipelineCycle will be lowered as:
   {stageA}, {stageA_recompute, stageB}, ipuCopies

This is represented in the IR as:
   {StageA} ipuCopies {stageA_recompute, stageB}    (note: stageA_recompute might be implicit)

When inplacing operations in the IR, tensors in stageA that are consumed by
ipuCopies will not present any conflicts as the tensors will already have
been consumed before stageA_recompute where they will be modified.

However, in the lowered version there is a conflict as {stageA_recompute}
is executed before ipuCopies. As such there is a constraint that any Tensor
produced by recomputation and is consumed by an ipuCopy must not be
aliased.

Test Plan: Add failure case to `pipeline_user_recompute_test` (and actually add the test to the CI)

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T33960

Differential Revision: https://phabricator.sourcevertex.net/D40776