Change LossScaleUpdate op to also take the inverse scale tensor as an input, and output an aliased version of it
Summary:
In popart's optimizer implementation, the are a number of 'atomic scalar' hyperparameters that the user can control:
`{mm, dm, wd, lr, ls, vs, rf, af}`

In the actual IR, loss scaling (`ls`) is represented as an atomic scalar. But the rest have been merged together as 'compound scalars' to reduce the required computation in the optimizer implementation. For example, for the SGD0 optimizer, there are only two compond scalars:
```
// Compound scalars for the case where there is no gradient accumulation and no
// momentum (SGD0):
//
//  - weightDecayScaleFactor0 (wdsf0) =
//      1 - lr * (1 - dm) * wd
//
//  - scaledLearningRate0 (slr0) =
//      lr *  ( 1 - dm) / ls
```
See here that the tensor `slr0` contains the operation that de-scales the gradient by diving by `ls`. So the `LossScaleUpdateOp` needs to update the `ls` tensor, and the `slr0` tensor (or whatever tensor contains the inverse `ls` term) by the inverse of the update factor.

This diff adds support for the second input/output of the LossScaleUpdateOp.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T32731

Differential Revision: https://phabricator.sourcevertex.net/D41048