Propagate target options to multi-replica target
Summary:
For multi-replica distribution, we create a compilation
"super-target" that covers all the IPUs in the system. Use
the new Poplar API added in D40955 to make sure that this
new target has all the same options as the original target.

This will be especially important once the `gatewayMode`
target option is added in D39723.

Equivalent diff for TensorFlow: D41068.

Ref T34794.

Reviewers: ismailp, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: ismailp, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T34794

Differential Revision: https://phabricator.sourcevertex.net/D41112