Update and re-run gen_python_docs
Summary:
The python docs haven't been updated to reflect some of the docs changes we have made.

Also the .sh file needs to be updated to remove deleted files.

This diff includes the update to the sh file and a re-run of it.

Test Plan: If it compiles on buildbots, then all is good.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Maniphest Tasks: T34834

Differential Revision: https://phabricator.sourcevertex.net/D41170