Test int8 host stream support
Summary:
Depends on D40996.
This already effectively adds int8 host stream support, as
we do not explicitly forbid them anywhere. We would error in
`popx::popType(DataType::INT8)` instead of returning
`poplar::CHAR`.

Thus, this diff mostly just adds tests.

However, there are two extra case to cover:

1) Pipelining+recomp
When pipelining with recomputation, if an input tensor is recomputed, it
gets stashed. Since we now support int8 inputs, we must support int8 stash
and restores. Unfortunately, these are implemented using
dynamic slicing ops, so a workaround is needed:

We cannot do a dynamic update based on tensor stashIndex, but we can do a
dynamic switch-case on stashIndex. There are stashSize cases, with each
case i being the program that should be run if stashIndex has the value i.
We have thus "unrolled" the dynamic update in a way.

As part of the implementation for restores, we factor out the common code
between RestoreInplaceOpx and RestoreOpx into RestoreBaseOpx. Also, we
make RestoreOp::getStashSize() a const method.

2) Synthetic data for an int8 input.
Extra support needed in functions:
a) `util.cpp:convertFloatToDataType` supports int8
b) `IrLowering::setInitTensorValTask` supports int8

Re a): The way `SyntheticDataMode.RandomNormal` works is that we generate
a normal distribution of floats then cast to the correct type. This diff changes this
behaviour to first `std::roundf` the data before casting.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T33689

Differential Revision: https://phabricator.sourcevertex.net/D41380