Disallow inplacing on an output of a recomputation
Summary:
In T35025 , an output of a groupnorm was being inplaced by a relu.

When it came to recompute, the batchnorm re-ran, which overwrote the output tensor of the norm, which in turn was used when computing the grad of a later conv. This caused an incorrect backward calculation. It is unclear why the conv grad was computed after the norm recompute, but investigation is ongoing.

See:

{F1282304}
{F1282305}
{F1282307}

We also suspect there is an issue w.r.t implicit recomputation not respecting topological constraints imposed by inplacing. Though we need to investigate further.

 (This issue was also present when using batchnorm - there was a lot of noise w.r.t batch norm, this may have been an unrelated issue.)

Fix for issue where recomputed ops affect the input to the backwards pass.

Test Plan: Let buildbots run.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, matthewha, davidn, fabiant, #popart

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, #popart

Subscribers: jamesn

Maniphest Tasks: T35025

Differential Revision: https://phabricator.sourcevertex.net/D42061