Added auto loss scaling transform with bare-bones support, enabled by a session option, and some tests
Summary:
Part 1 of automatic loss scaling (ALS). The motivation for this transform, and a description of it is at the top of automaticlossscaling.hpp.

So far it has very limited compatibility with popart's various execution modes. It is not compatible with:
- sharding
- pipelining
- graph replication
- gradient accumulation
- non-SGD optimizers

Tests have been added that an appropriate error is thrown if any of these are enabled alongside ALS. Adding support for these execution modes will follow in later diffs, but have been omitted to keep this diff straight forward (basically just an IR transform).

Fixes T32733
Fixes T32731

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: badreddine, matthewha

Maniphest Tasks: T32731, T32733

Differential Revision: https://phabricator.sourcevertex.net/D40538