Update python docs and add missing docstrings where appropriate.
Summary:
This adds some detail to the session options, in particular for the enums. Unfortunately as it is currently bound, the session options do not specify the enums that they use. I am unsure how to get this to be correctly documented.

I also re-ran the gen_python_docs.sh file to update any missing documentation.

This is the docs that are generated:

{F1233581}

Test Plan: It builds correctly for me.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Subscribers: jamiep

Maniphest Tasks: T33792

Differential Revision: https://phabricator.sourcevertex.net/D40047