Add OptimizerFromHostFragment to ExecutionContext enum, and set the settings.executionContexts of optimizer-tensor-copying operations to this
Summary:
When you have an IpuCopy op in the graph that copies an optimizer tensor, e.g.
```
learning_rate ---> IpuCopy ---> learning_rate_c0
   vg0                              vg1
```
this IpuCopy op will not be executed as part of the main graph (i.e. in the `ExecutionContext::Normal` execution context). In fact, it will not be executed in any context that is run with the call to `session.run`. It will only be executed with a call to `session.updateOptimizerFromHost`. There is a devicex program fragment for these ops, called `OptimizerFromHost`, but for some reason the `ExecutionContext` enum class does not cover this. I believe there is no good reason for this omission.

Here I have added the context and set it where appropriate.

This is a precursor to the work for T33956, for which is desirable to be able to control the execution context of optimizer-tensor-copying ops depending on whether automatic loss scaling is enabled

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T32731

Differential Revision: https://phabricator.sourcevertex.net/D42221