Adding unittest and mocking support.
Summary:
With this diff I am proposing we add a new type of test in a new folder, `popart/tests/unittests`. These new kind of tests are 'proper' unit tests (not integration tests) that meet the following criteria:

* They run extremely quickly (<< 0.1sec).
* They test 1 class' code in isolation (or at worst a bunch of highly coupled classes, but try and avoid this).

Unittests should not call poplibs or poplar, or instantiate other PopART classes, or use the python interface, or depend on time progression. If we cannot test a class in this way some refactoring may be required to make this possible.

I am not suggesting this as an alternative to the existing test method as integration testing is also valuable, but I am suggesting that aiming to write this type of unittest for new code we write is probably always a good idea.

Note that a mocking framework, trompeloeil, has been added as a dependency of PopART so that it is 'easy' to control dependencies of classes under test.

Test Plan:
  * Check existing tests run on diff builder.
  * Check new tests runs there, also.

Reviewers: jamesn, evan, shirazb, george, matthewh, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, #popart

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, #popart

Subscribers: matthewha

Maniphest Tasks: T36181

Differential Revision: https://phabricator.sourcevertex.net/D42111