Refactor IfOp
Summary:
Part of T35979, deprecate Op::setup.
Added a new constructor to `IfOp` that omits `inputIds` and changed the factory function to connect these inputs.
The old constructor is still in place and the inputs will still be connected in `IfOp::setup`, but this has been deprecated and can be removed after the next release.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Maniphest Tasks: T35979

Differential Revision: https://phabricator.sourcevertex.net/D42869