Remove getting the tile mapping from the public API
Summary:
Fixes T36337.

This is quite controversial because it skips the deprecation period for removing the public API however I think we need to do this so putting it up for review. If we don't do this we could change it to return an empty map always and deprecate the API at the same time, but if we don't do either one of those we cannot insert popIR into popART in the 2.1 timeframe.

Reviewers: jamesn, fabiant, george, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: jamesn, george, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: evan, #software_-_popir

Maniphest Tasks: T36337

Differential Revision: https://phabricator.sourcevertex.net/D43030